SoundList implementation
First commit for this feature. Sounds can now be organized into lists.
Currently, the only way to actually put a sound into a new list is by doing the
SQL insert yourself. Next commit will have this implemented. I just wanted to
commit this now because it was a lot of work :)